Output Correct primary MediaObject for the configured UploadableField annotated properties

+++ src/Uploadable/UploadableHelper.php
@@ -13,9 +13,13 @@ declare(strict_types=1);
 
 namespace Silverback\ApiComponentsBundle\Uploadable;
 
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Persistence\ManagerRegistry;
+use League\Flysystem\Filesystem;
+use League\Flysystem\UnableToReadFile;
 use Silverback\ApiComponentsBundle\Annotation\UploadableField;
 use Silverback\ApiComponentsBundle\AnnotationReader\UploadableAnnotationReader;
+use Silverback\ApiComponentsBundle\Factory\Uploadable\MediaObjectFactory;
 use Silverback\ApiComponentsBundle\Flysystem\FilesystemProvider;
 use Silverback\ApiComponentsBundle\Model\Uploadable\UploadedBase64EncodedFile;
 use Silverback\ApiComponentsBundle\Utility\ClassMetadataTrait;
@@ -31,15 +35,18 @@ class UploadableHelper
 
     private UploadableAnnotationReader $annotationReader;
     private FilesystemProvider $filesystemProvider;
+    private MediaObjectFactory $mediaObjectFactory;
 
     public function __construct(
         ManagerRegistry $registry,
         UploadableAnnotationReader $annotationReader,
-        FilesystemProvider $filesystemProvider
+        FilesystemProvider $filesystemProvider,
+        MediaObjectFactory $mediaObjectFactory
     ) {
         $this->initRegistry($registry);
         $this->annotationReader = $annotationReader;
         $this->filesystemProvider = $filesystemProvider;
+        $this->mediaObjectFactory = $mediaObjectFactory;
     }
 
     public function setUploadedFilesFromFileBag(object $object, FileBag $fileBag): void
@@ -79,7 +86,7 @@ class UploadableHelper
                 continue;
             }
 
-            $filesystem = $this->filesystemProvider->getFilesystem($fieldConfiguration->adapter);
+            $filesystem = $this->getFilesystemFromFieldConfiguration($fieldConfiguration);
 
             $path = $fieldConfiguration->prefix ?? '';
             $path .= $file->getFilename();
@@ -92,12 +99,60 @@ class UploadableHelper
         }
     }
 
+    public function deleteFiles(object $object): void
+    {
+        $classMetadata = $this->getClassMetadata($object);
+
+        $configuredProperties = $this->annotationReader->getConfiguredProperties($object, true, true);
+        foreach ($configuredProperties as $fileProperty => $fieldConfiguration) {
+            $currentFilepath = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
+            if ($currentFilepath) {
+                $this->removeFilepath($object, $fieldConfiguration);
+            }
+        }
+    }
+
+    public function getMediaObjects(object $object): ?ArrayCollection
+    {
+        $collection = new ArrayCollection();
+        $classMetadata = $this->getClassMetadata($object);
+
+        $configuredProperties = $this->annotationReader->getConfiguredProperties($object, true, true);
+        foreach ($configuredProperties as $fileProperty => $fieldConfiguration) {
+            $propertyMediaObjects = [];
+            $filesystem = $this->getFilesystemFromFieldConfiguration($fieldConfiguration);
+            $filename = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
+            if (!$filename) {
+                continue;
+            }
+
+            // Populate the primary MediaObject
+            try {
+                $propertyMediaObjects[] = $this->mediaObjectFactory->create($object, $filesystem, $filename);
+            } catch (UnableToReadFile $exception) {
+            }
+
+            // Todo: populate media object for configured ImagineBundle filters
+
+            $collection->set($fieldConfiguration->property, $propertyMediaObjects);
+        }
+
+        return $collection->count() ? $collection : null;
+    }
+
+    private function getFilesystemFromFieldConfiguration(UploadableField $fieldConfiguration): Filesystem
+    {
+        return $this->filesystemProvider->getFilesystem($fieldConfiguration->adapter);
+    }
+
     private function removeFilepath(object $object, UploadableField $fieldConfiguration): void
     {
         $classMetadata = $this->getClassMetadata($object);
 
         $filesystem = $this->filesystemProvider->getFilesystem($fieldConfiguration->adapter);
         $currentFilepath = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
-        $filesystem->delete($currentFilepath);
+        if ($filesystem->fileExists($currentFilepath)) {
+            $filesystem->delete($currentFilepath);
+        }
     }
 }
